infrastructure/security/epic-7-story-2/functions/requirements.txt
google-cloud-secret-manager==2.16.4
google-cloud-monitoring==2.16.0
google-cloud-iam==2.12.1
google-cloud-pubsub==2.18.4
requests==2.31.0